LPS-15051 Generated
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@71694 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/service/com/liferay/knowledgebase/model/ArticleClp.java
@@ -167,11 +167,11 @@ public class ArticleClp extends BaseModelImpl<Article> implements Article {
 		_description = description;
 	}
 
-	public int getPriority() {
+	public long getPriority() {
 		return _priority;
 	}
 
-	public void setPriority(int priority) {
+	public void setPriority(long priority) {
 		_priority = priority;
 	}
 
@@ -514,7 +514,7 @@ public class ArticleClp extends BaseModelImpl<Article> implements Article {
 	private String _title;
 	private String _content;
 	private String _description;
-	private int _priority;
+	private long _priority;
 	private int _latest;
 	private int _status;
 	private long _statusByUserId;
